test/dbt_integration_tests/dbt-gitlab/pyproject.toml
[tool.poetry]
name = "analytics"
version = "1.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10.3"
dbt-snowflake = "1.8.3"
dbt-core= "1.8.0"
dbt-adapters = "1.3.2"
gitlabdata = "0.3.28"
fire = "0.4.0"
loguru = "0.6.0"
numpy = "1.26.4"
pandas = "1.4.2"
SQLAlchemy = "1.4.46"
simple-dependency-resolver = "0.1.1"
requests = "^2.31.0"
sqlfluff = "3.1.0"
sqlfluff-templater-dbt = "3.1.0"
